Change dial pad to show chooser onStart instead of onCreate.
This fixes the problem where the chooser is not shown when the dialpad is
started and pushed to recents before a call comes in.  When the call comes
and dial pad is selected from recents, the dialpad is shown.  Instead, we
should be showing the chooser so the user has the option of getting back
into the call.

Bug: 5761141
Change-Id: I315117c6607dbb71e5359cf98785591d20c21881