Remove hard coded PeopleActivity.class from VCard UI.
The PeopleActivity was hard coded as the target activity when clicking on
a vcard processed notification.  Since the vcard export can be triggered
from both the dialtacts UI and the people UI, it's strange to always go
back to the people UI.

Change vcard UI classes to accept the calling activity. This way we can
send the user back to the activity that originally triggered the vcard
export.

Bug: 6993891
Change-Id: Ice978acb8244aacfcdc94c7da0ccc1e43bada54c