Add a "No contacts" description to PhoneFavoriteFragment
Add an additional "No contacts" view to the sub-header view
returned by PhoneFavoriteMergedAdapter. This is required because
the empty view functionality provided by a simple list view does
not work because PhoneFavoriteMergedAdapter is a combination of
two adapters. This "No contacts" view provides the illusion of
an empty view for the 2nd adapter that displays all contacts
without phone numbers, even if the first adapter does contain
contacts.

Bug: 5522729

Change-Id: I7dcc4cd7b4b2b28fefea4be384de40318654e165