Remove smart dialing content observer
The content observer doesn't help to detect changes that happen
outside the application (e.g. user modifies a contact in the People
app), and seems to be called multiple times while in the dialer for
no reason, causing multiple unnecessary recaches. Instead, just force
a recache in onResume upon startup to refresh the cache.

Bug 8527847

Change-Id: I52f5c435ad573ce8b195d7b7828f19501e068c60