Animation for smart dialing suggestions
Suggestions now appear with a fade in and slide up animation.
Suggestions vanish with a fade out and slide down animation.
If a suggestion is moved into the middle, it slides to the left/right
as appropriate.

Change the layout containing suggestions to a LinearLayout, in order to
better support animations.

Renamed SmartDialAdapter to SmartDialController, and also refactored
it to handle entries for a LinearLayout instead of a GridView, as well
as adding animation support and view management.

Use null object pattern in SmartDialEntry to better handle null entries.

Start displaying suggestions on the first digit entered.

Bug 8840240

Change-Id: If4e16006c0b36d2244434e0b2d8f3d3b997b0ad2