Add call shortcuts  to phone number list adapters
* Add a new class DialerPhoneNumberListAdapter, which both
RegularSearchListADapter and SmartDialNumberListAdapter extend.
DialerPhoneNumberListAdapter allows the addition of pre-defined call
shortcuts at the end of the phone number list that can be toggled
on or off.

* For the SmartDialSearchFragment, the only shortcut that is available is
always the Add number to contacts shortcut.

* For the RegularSearchFragment, if the user enters a string that contains
all dialable numbers, it is treated as a phone number and the call directly
and add number to contacts shortcuts are enabled. Otherwise, it is treated
as a name, and only the add contact with a new name shortcut is enabled.

* Add a intent that allows the user to directly create a new contact
that has the name field pre-populated. This intent is used if the user
enters input that looks like a name in the search view.

Bug: 10339630
Change-Id: I2ae757ce505d85a8780d28d89e09fb7084c773b1