Update patch set 4
Patch Set 4:

After thinking about this more, I'm not convinced of the approach yet. The NameToNumber stuff essentially doesn't do latinization (or rather romanization, my wording was off there), but still converts the name to numbers. With the special classes this works for chinese and korean, but not e.g. for russian, thai, or any other language that doesn't use latin characters or characters which can be trivially converted.

My gut feeling is that we need a general romanization routine (the pinyin one does romanization, the korean one doesn't).

Patch-set: 4