Update patch set 6
Patch Set 6:

The problem - as I see it - is that the smart dialer code simply works differently from the T9 dialer up to CM 10.1:

- The CM T9 dialer converted the display name into a series of digits (mapping characters to single digits) and afterwards compared the conversion result to the inputted digits. Conversion from characters to digits was done by a language dependant table.
- The new code tokenizes the display name prior to do any matching. Tokenization is done by splitting at characters that are neither digits nor roman characters. Each token is converted to a string of digits by a hardcoded latin-character-to-digit table.

The new code so only supports roman characters. You avoid that problem by converting the chinese and korean names to digits again, which is why it works. However, I don't think it's the correct solution because
- it still is broken for cyrillic and any other non-roman alphabet
- the tokenization still doesn't work for any non-roman alphabet name
- putting a second character-to-digit conversion on top of another one doesn't look like a correct solution

So in essence, IMHO we have to do one of two things: either
- extend the AOSP conversion routine to deal with non-roman characters or
- do a generic name romanization and match on the romanized name

Neither is terribly easy to do, though :(

Patch-set: 6