Fix missed calls not updating properly
This fix updates the new/old state of missed calls in the call log
properly once more.

Instead of using this flag to track whether or not to show the call
shortcut, we store a date in Dialer's shared preferences to allow
us to only show call shortcuts newer than a certain date.

Bug: 10780687
Change-Id: Icc03992a65512a85dc6673c431ded34c73b6ba2b