Fix accessibility issues with dialpad
Each dialpad key needs to manually have the "focusable" flag set so that
hovering and lift-to-type works properly.

The edittext containing digits needs to receive focus at least once in
order for it to report sendAccessibilityEventTypeViewTextChanged type events.
In JB-MR2, the ViewPager would automatically give focus to the dialpad
fragment when it was swiped into view, but now we have to manually give it
focus ourselves when the dialpad fragment is swiped into view.

Bug: 11067558
Change-Id: I56a1891b12d5ec7771a37e1f7c361242ebfe028d