update dialpad state more often.
Noticed while testing proximity sensor that we do not update the dialpad
state sufficiently. We have increased the amount of places in the code
where the dialpad comes up since we first introduced the proximity code
and the callbacks wheren't updated.

This manifested in situations where the proximity sensor was off at time
when it should have been on.
Repro: be in a call, answer a second call (call waiting), bring up
dialpad, hangup second call.  The dialpad would go away but the prox
sensor still thought it was up.

This change moves the callback to proximity sensor to the exact
place where dialpad is shown/hidden in order to catch all instances.

bug:11006867
Change-Id: I7651b4769d035b9bc68d226da3ffb3aea8986e58