Load contact in CallDetailActivity and add it to Add contact intent
Use a ContactLoader to load the contact data stored in the json-encoded
URI asynchronously, and prepopulate the add contact intent with the
contact data when the contact data is loaded.

This fixes the problem of contact details (i.e. name, photo, address)
not being pre-populated when adding a contact via a call log entry that
was created by calling a contact from nearby places search.

Bug: 11294679
Change-Id: I39a0d8ae40e80b350687a43be8ece6b32c88062a