Fix bug where resuming CallDetailActivity loses the add contact intent
getLoaderManager.initLoader() performs the onLoadFinish callback
immediately if CallDetailActivity is being resumed from a previously
paused state. This causes bindContactPhotoAction to be called
with a null intent later on causing the add contact button to be no
longer clickable.

To fix this, if we know that we will receive a loader call back
that correctly binds the intent later on (or immediately), we should
skip the initial bind.

Bug: 11588776
Change-Id: I9407a88ec8bbde303109eb14496e01cf6c347c8f