Update patch set 1
Patch Set 1:

I didn't know if there was a proprietary dialer only for nexus devices or not.  As far as I could tell the only different I could visibly see was the icons so I assumed it was an oversight made by either Google or CM.  I guess I could decompile dialer.apk from Google and compare it to the CM version to see if there really is a proprietary vs aosp  version.  I just thought they were one in the same and the only thing that was proprietary was potentially the Google Experience Launcher.....

Patch-set: 1