Update patch set 1
Patch Set 1:

I guess a way to test how proprietary Google considers them could be to submit this change to AOSP and see if they merge it. If they do, there shouldn't be any issue. Google forget to update small things like this all the time so the lack of them in AOSP isn't necessarily indicative of an intentional omission.

Patch-set: 1