Add incoming-during-hold case to fullscreen Notification code.
We use fullScreen notifications to bring up the incoming call screen
when it is not already being shown.  Android framework ignores
fullScreen property if the same notification ID already exists. To fix
that we cancel the notification right before setting it to full screen.

We only used to do that for the CALL_WAITING state since that when it
mattered.  However, if the current call is on hold, a new call comes in
as INCOMING instead of CALL_WAITING so this change adds that condition
as well.

Also fixed two other minor pieces of code which only checked INCOMING
and not CALL_WAITING.  One for an icon, the other used to acquiring
location as part of the contact info (we only do so for incoming calls).

bug:11578080
Change-Id: Id010a6848cd1ca06d68e8405041565bc008a781e