Fix possible NPE in DialerDatabaseHelper
In certain cases (third party contact backup solutions?) it is possible
for raw contacts to end up with null display names, which the Dialer
was crashing on. This fix performs a null check and inserts the default
(No name) contact name if we find that the display name is null.

Bug: 11750613
Change-Id: Ifbaf2861534e6eaf8a86d50d584d11a9fa38b65f