Update patch set 1
Patch Set 1: Code-Review+1

I agree with Bjorn. All that people see is an icon and think "Hey I thought CM11 is KitKat, why do I have the old JellyBean icon!".

All KitKat themes out there change the dialer icon, because that is what people want, the same icons as on the Nexus 5. They know that changing an icon does not give them the search functionality.

All the other icons are changed to the KitKat ones, would be weird not to update this one too.

Patch-set: 1
Label: Code-Review=+1