Temporarily stop UI refresh during drag and drop
If the user is in the middle of a drag and drop operation, temporarily
stop UI refreshes when a new cursor is delivered, to prevent UI
inconsistencies.

This CL only fixes the dragging/dropping case. Swiping is still buggy
(the fix is more complicated/risky), and swiping might be removed
altogether in the future.

Bug: 11119329
Change-Id: I2939c915d4eb5d8f9de19a369b21c6c98806ba94
(cherry picked from commit 670fd3f59160bdb81f38b6be2f176ba233915686)