Always call through to super.onBackPressed in DialtactsActivity
moveTaskToBack is causing crashes in the Dialer when the back button is pressed.
This was legacy code that was unchanged in the new dialer. We should probably
just call through to super.onBackPressed regardless of whether or not the task
was at the top of the stack, to behave the same (navigation wise) as other apps.

Bug: 12015851
Change-Id: Ic0c3efb30b3874ad892b96a9fadd524313cfb782