Update patch set 4
Patch Set 4:

That's because you didn't set a transparency on the main background. background_dialpad is still plain #eeeeee.

The transparency for the digits part also needs to be much lower so that you can actually see through both layers. Something like 80 works better.

Setting a color background on the backspace key also has the unfortunate effect of removing touch feedback from the button. That should be fixed with a selector resource.

Patch-set: 4