Update patch set 4
Patch Set 4:

Isn't that the point? Right now, having the dialer panel slide out with the exact same background color as the contacts list removes all depth from it. There's no visual indication it's an overlay. That's probably why it's transparent in the AOSP version too. I think a big reason for the dark color scheme in AOSP is to visually separate it from the rest of the dialer: right now that isn't happening with this change.

Also, since the number input view is overlaid on top of the main panel background, it needs to be much more transparent to be actually see-through (it needs to show through both its own white background and the dialer panel itself). For this reason, it should have a transparency value of, say, 80 or something close to that.

Patch-set: 4