Update patch set 9
Patch Set 9:

I submitted a separate commit here (http://review.cyanogenmod.org/#/c/57190/) that is more focused on restoring the style changes and color redirections from cm 10.0-10.2 so that the Dialer can be fully themed with the Theme Manager. Out of curiosity, would it make sense to add those changes into this commit as well, or keep them as 2 separate commits? The only reason I ask is that there are a few colors whose values would change depending on whether you're looking at the current merged code, or this commit.

Patch-set: 9