Update patch set 9
Patch Set 9: Code-Review-1

Actually, making this new UI to dialer, will break the logic of the blue icon dialer has. Also, afaik, CyanogenMod does not change Google AOSP UI too much often, it improves them (on code), so we should focus on this http://review.cyanogenmod.org/#/c/57190/ giving the dialer the option to choose a theme, instead of chaning its UI. Then we could create a theme to mod this dialer and post it on XDA or other websites.
I'm not telling I do not like the UI of new one, I'm telling it breaks logics (with actual icon) and it could be a better alternative using themes. Again, my opinion, nice work though :)

Patch-set: 9