Update patch set 9
Patch Set 9:

There is no reason we can't do both, this patch and theme engine compatibility. 

The discussion of what we should focus on is not relevant, there are plenty of developers/contributors looking at various aspects of the software. Whether that means diverging from what AOSP provides or not isn't the issue - consistency is. 

Kitkat can move to the white/flat UI, but at least do it consistently - as it stands the darker dialer clashes with the rest of the app and the contacts app. That is more than enough justification.

Patch-set: 9