Update patch set 10
Patch Set 10:

Out of curiosity, since there seems to be quite a bit of  differences in opinion on the level of transparency, would it be possible to blur the background behind the dialpad such that you can retain the transparency without the items behind the dialpad being visually distracting or making it difficult to read? What I had in mind is something similar to the blurred lockscreen background on AOSPA and PAC-man. The commit for that can be found here: https://gerrit.paranoidandroid.co/#/c/382/

Screenshot here: https://db.tt/ZDNoHjy3

One thing I'm not sure about is how one might go about limiting the blur to just the area behind the dialpad without blurring the other contacts/favorites that are shown above the dialpad in the top 1/3 of the screen.

Patch-set: 10
Reviewer: m4570d0n <4381@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0