Update patch set 1
Patch Set 1:

lol yes it is a gray zone indeed. But latest now that CM got attention on it it cannot be merged anymore :p

I will look to work that with Xiao-Long out to have a open source implementation which is an own one but based on the idea.

Patch-set: 1