Fix contact tile horizontal animations when swiping shortcut card
Instead of using the contact tile's position in the listview to
determine what kind of horizontal animations it should undergo, use
its position in the contact tile adapter. This ensures that its
position remains the same if the shortcut card is swiped away,
but is still calculated correctly if its actual position within
the contact tile adapter changes.

Bug: 13341497

Change-Id: Ia5d4bad1e01a4e47a3fcd934dd580c76ca183a20