Update patch set 4
Patch Set 4:

Jeff: The reason that happens is because only the in-call screen and call history/log will use reverse lookup. If you type in a number for, say, McDonald's, the nearby places feature won't show "McDonald's". (It works the other way around). The assumption is that if you're trying to dial a number, you probably already know what it is.

Patch-set: 4