Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

We should not merge this commit anyway, weird stuff is going on with the arrays (see inline comment)

About the dialer letters: cm_extra is not picked up by Crowdin, so I don't think adding it there is a good idea. Need to find a better solution, or this problem will happen on every sync.

Patch-set: 3