Move call shortcut cards to their own list
This CL migrates the logic that queries for and displays call
shortcut cards to ListsFragment so that they can be displayed
above the main contacts viewpager.

* Add a ListView above the ViewPager in ListsFragment

* Implement CallLogQueryHandler.Listener and CallLogAdapter.CallFetcher
in ListsFragment, and delete that corresponding logic in
PhoneFavoritesFragment.

* Use animateLayoutChanges on the parent LinearLayout hosting shortcard
cards so that we don't have to handle animations ourselves when a card
is swiped away.

* Make the minimum set of changes to PhoneFavoritesMergedAdapter so that
only call shortcut cards are displayed (no more menu, teaser, etc).
More significant changes to actually delete unnecessary logic is upcoming
in a separate CL.

Bug: 13963734

Change-Id: I55cdc53a9311b1bb78422aab9d174e860596b997