Disable quickcontactbadge in AllContactsFragment
Disable the quick contacts badge in AllContactsFragment and just
use the regular photo ImageView so that contact photos can be made
round and have the overlay removed without affecting other apps
that use the QuickContactsBadge (e.g. Gmail).

Also tag the view with the contactUri so that it can be used to
bring up the QuickContactActivity when clicked.

Bug: 13747208

Change-Id: I02288fd64d133d75c7b5e635d9b450f0b2cf7673