Always show the call state text view.
Before this was flashing in and out because of the change in
visibility, which was causing some shifting and UI jank. Now
the call state label will always take up space in the layout.

Also, moved where the elapsed time view is so that it is
properly aligned to mesh with new changes.

Change-Id: Id8b395b18ebc1038310721d878622c6b0c15c295