Cleanup dialpad show/hide logic.
This is the first step in cleaning up the DialtactsActivity logic to
make it easier to transition between different states, and handle
the states in between (hopefully).

Originally, I thought to also block show/hide actions if there was
already a dialpad transition in progress, but in the end I did not
do this because it made the UI seem unresponsive.

- Delete custom predraw listener in DialpadFragment; it doesn't
appear to be used (?) and made the animation logic more confusing.
- Replace the fragment custom animations with our own animations,
which seem to perform better.
- Changed the animation xmls from objectAnimators to translates.
- Replaced isDialpadShowing with uses of mIsDialpadShown.
- Reorder show/hide logic. Now it is less parallel, but better
reflects the ordering/nature of fragment transactions.
- Reorder some methods to group them more logically.
- Pull out a helper method for updating the search fragment's
yTranslation.
- Pull out a helper method for updating the floating action bar.
- Delete predraw method in DialpadFragment; doesn't appear to be
used and caused.

Bug: 14900155
Change-Id: I768324b2293622ea43974f7f566eec9b11fb0d4e