Show elapsed time in InCallUI.
Before, this was getting owned at the end of animation, for some
reason (perhaps the visibility was being set improperly?). In any
case, the newer AnimUtils in contacts common handles this properly,
so here we delete the hide/show helpers in AnimationUtils which
are now no longer needed and use these more common utilities instead.

Bug: 15281475
Change-Id: I86558d51b14605fff70945c3a97df05ddc45f1ca