Set call log item view alpha to 1 after expand animation end.
This bug is on the inexplicable side, but this seems like it fixes it.
Without this, after navigating out of the activity and then back, it
appears taht the the action view's alpha defaults back to the value
being set at the start of the expand animation (0), making the view
invisible.

Bug: 15275605
Change-Id: I8173f305569e2e8f5e13f34235922b830dd4e532