DO NOT MERGE. Styling search to match redlines.
- This conflicts with some of Sai's changes, so will manually
cherrypick this CL back to master..
- Shift arrow/close icons.
- Use 20sp font in expanded search (causes baseline to shift down).
- Update action bar movement to use dialpad's interpolators and
durations.
- Smoothed out animations. Icons in the view being faded out are now
instantaneously set to GONE. The margin animator when going from
collapsed to expanded starts at 0.8f now, so that the searchbox does
not shrink momentarily when switched to the 9-patch, because of the
shadow.
- Fix clipping because of 9-patch; refactored logic for updating
the position and padding from the Activity into SearchFragment. This
was motivated in part because we want a padding at the top of this
list for the first item, but when the dialpad is shown we do not
want that padding.

Bug: 15343931
Change-Id: I7c3a96dd295c58c6955e6bd3e9d8cda151de4c3f