DO NOT MERGE. Styling search to match redlines.
- Cherrypicking changes back to master. Mostly, this involved
transferring changes dealing with padding and the translation values.
- Shift arrow/close icons.
- Use 20sp font in expanded search (causes baseline to shift down).
- Update action bar movement to use dialpad's interpolators and
durations.
- Smoothed out animations. Icons in the view being faded out are now
instantaneously set to GONE. The margin animator when going from
collapsed to expanded starts at 0.8f now, so that the searchbox does
not shrink momentarily when switched to the 9-patch, because of the
shadow.
- Fix clipping because of 9-patch; refactored logic for updating
the position and padding from the Activity into SearchFragment. This
was motivated in part because we want a padding at the top of this
list for the first item, but when the dialpad is shown we do not
want that padding.

Bug: 15343931

Conflicts:
	res/layout/search_edittext.xml
	res/values/dimens.xml
	src/com/android/dialer/DialtactsActivity.java
	src/com/android/dialer/list/SearchFragment.java

Change-Id: I8519e7e2ef5c75c0f170ff61ed22f44f95e07f74