Always show options menu in Dialer dialpad.
We disable options if there are no digits in the text field. The
PopUpMenu does not reflect this visually right now though. This
also fixes bad centering of the "dial to add call" text, because
now the text field will have equal left/right margins from the
edges of the screen..

Bug: 15453278
Change-Id: I6f4a7a597ed289fb1b59c9fcac261d151144b642