Update patch set 1
Patch Set 1:

Ricardo, this patch is a workarround. From what i've seen, the problem is that the button is not attached to the listview (but visible) and the click events are enqueued but not dispatched. This is caused by "mContactTileFrame.setLayoutParams(params);" code that is set when there is not favorites. This caused that the listview changed its size and i think the view is dettached but no attached until other event force it. I tried invalidates, requestRelayout, posting,..., but nothing seens to have effect. The only way i found is to have the height of the view inmutable, and redesign the layout to fit the nofav panel with change the contacts panel height. Not sure if you know a way to get the view reattached again. I think is a bug in the ListView with multiples view types.

Patch-set: 1