Update patch set 1
Patch Set 1:

Ricardo, sorry for pointing you in the wrong direction. The requestLayout race condition it's not the cause of this bug. Is just the explanation for why it works sometimes and doesn't works others. That is: a requestLayout event after the detach of the view makes the view to by reatached again.

And now I'm almost sure that the detach event should't be called in this case. This is happening because the active view is being promoted to the scrap views array in abslistview and remove and detached by pruneScrapViews. Or at least this is my current theory (i have my mind near to blow up :).

I need to figure out it is caused by a bug in the abslistview or in the dialer adapter. That is my homeworks for today ;).

Patch-set: 1