Clear fragments properly when search term is erased.
Specifically, in the scenario in the bug the dialpad was not shown
but we actually wanted to clear the speed dial fragment, not the
regular search fragment. Change this to affect the visible fragment,
rather than whether or not the dialpad is shown.

This onClick pathway is no longer used too, so deleting.

Bug: 15516704
Change-Id: Icd4419e0746ce0e2bcb9c9437dbac77d64bb8c98