Update patch set 1
Patch Set 1:

As this spinner is supposed to be shown in one of the three fragments of this activity only, I figured it's easier and less error prone to put it into a fragment specific menu item than to dynamically change the navigation mode depending on the active fragment. It's probably possible to implement it either way, but the visual outcome won't differ much anyway, will it?

Patch-set: 1