Simplify call-button-presenter update UI code.
I consolidated a lot of the if/then branches into booleans. This
means that buttons will be updated more frequently, but I think
it has the positive effect of making it easier to read, and is
considerably more concise.

I'm considering moving this code into a helper to address updates
for audio calls (vs. video calls), and thought to clean this up
on its own before doing that.

Bug: 16014280
Change-Id: I5ef53ce654091f033936233b7368b54b45cc3542