Status labels for video calls.
- Display the appropriate status labels for video calls, mostly
depending on whether one is in a video call, and whether one
has requested video and is waiting for a response, or a video
request has failed.
- Add a handler to the CallCardPresenter, so we can show an error
message in the fail case, but only show the message for a time.
- Add video icon in layout.
- Renamed CallVideoClientNotifier to reflect newer nomenclature.
- Add session modification state information on InCallUI call. This
helps us track state to know what strings to display on the card.

Bug: 16013340
Change-Id: Ib2bf84d93a05664adbf7fe838848b7d7b54a8254