Update patch set 5
Patch Set 5:

(1 comment)

I wonder whether the fact that recordings are present should be indicated in the call log itself. It'll get a bit tough to do a query with decent performance for that, though.

And BTW, is the playback icon really colored in Holo light (black with 70% opacity)?

Patch-set: 5