Update patch set 8
Patch Set 8:

Chris, I agree it would be nice to have an integrated way to delete recordings.  But I would prefer to do that as a subsequent change rather than as a part of this one.  If the feature turns out to be popular, there are a lot of things we can do to improve it (and I would put deletion near the top of the list)

Patch-set: 8