Update patch set 1
Patch Set 1:

The transposed laws may vary between EU member states. In Portugal, for instance, you must be informed at the start of each and every call, give explicit consent, and the question&answer must be included in the actual recording.

In any case, like Ed mentioned, this isn't intended to be used yet and we're just trying to prevent the builds we distribute out of get.cm from accidentally including the feature in regions where it is forbidden (or of dubious legality). Distributing a ready-to-use consumer product (in the form of builds) with this capability may expose us to legal issues :)

Patch-set: 1