Update patch set 1
Patch Set 1:

i think you're misusing the term "illegal" here.
the law in some places around the world forbids making a recording completely, and in others allows it under any circumstances, or with some restrictions (e.g: both parties must be aware of a recording being made).

but it does not forbid any software from containing an ability to record. that would be ridiculous.
otherwise, recording devices would be banned, and you should also disable the voice recorder, because it may also be used to record a phone conversation, or someone's private conversation without their consent...

there are apps that do exactly this on the play store, sold in the u.s and eu, and they are not illegal (otherwise google would have removed them and banned their developers).
this ability is also included in skype (don't have it on my phone but my pc has the ability). it is up to the user to obey or disobey the law with the appropriate consequences...

just like it is not illegal to own a gun, but using it - not within a specified set of circumstances, constitutes a crime...

i don't see any reason this should not be enabled and available to everyone. has a lawyer actually told you this should be disabled (i highly doubt that)?

Patch-set: 1
Reviewer: Tom Keidar <6879@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0