Update patch set 1
Patch Set 1:

Excellent point Fabian.

If someone used a camera for illicit purposes, no one would think of suing Nikon or Canon, or Google if the camera used Android OS. Is the US/EU geo-blocking (which represents a significant part of the developed world) reasonable or overkill? And is this blocking in itself legal or discriminatory?

Patch-set: 1