Update patch set 1
Patch Set 1:

In http://review.cyanogenmod.org/#/c/69850/, Aug 12 1:23 PM, Abhisek stated "The legal items have already been vetted by lawyers. The onus to inform parties that the call is being recorded is on the participant(s) recording, not us.

The build time overlay prevents this from use unless the user explicitly builds with it, or in our case, the build is distributed in a region that sanctions such behavior."  
-----
Applicable laws apply to *THE USER*, not the device and not the software. Microsoft cannot be held responsible if someone uses a device running MS Windows to commit child/drug/endangered species trafficking crimes or other restricted or illegal activity. This "Fear of being sued" is baseless.

Patch-set: 1