Update patch set 1
Patch Set 1:

The end goal is to pickup the features required to be cmcc compatible, even if they are not things we would use ourselves.  One thing we know about China is that users rely on the dialpad more heavily, for example t9 search is the preferred way to search for contacts, so it doesn't surprise me that speed dial is something they would want.

Patch-set: 1