Update patch set 1
Patch Set 1:

Apparently (I had to look it up) these characters can be used to add a pause or wait into a numeric sequence which can be used to pre-program certain key presses for making it past automated phone menus.

http://www.androidpolice.com/2010/05/10/how-to-add-hard-wait-and-soft-2-3-sec-pauses-to-your-android-contacts/

Normally you would have to add a contact to do this but I suppose this change lets you do it directly from the dialpad.

Patch-set: 1