Call recording encoder/format choice 1/3
Adds a user option between WB AMR (the default) or HE-AAC for the encoding of call recording audio. Also changes the output format of the file appropriately (WB AMR now uses the WB_AMR output format and HE_AAC chooses the MPEG_4 output. File extensions are adjust accordingly as well).

This was in reponse to people complaining the quality of call recording audio is poor. It's noticibly better using HE-AAC, but filesizes are larger.

Change-Id: I95620bf944e18491652c716890396c3da4be70c4