Hide the FAB if the dialpad chooser is shown.
This bug was happening because in onHiddenChanged, the dialpad button
was being made to be shown. Now, only reveal the fab at that point if
the dialpad chooser is not visible.

There is some slight quirk... I would have thought that checking
!mDialpadChooser.isShown() should have been sufficient, but it
returns false at that point. The isDialpadChooserVisible check
works fine.

Bug: 17609779
Change-Id: I11eb3dc81a2849874202a3ea06e40c097b50ac9a