Update patch set 2
Patch Set 2:

The change to use the style looks good.  

Is it correct to assume that the activities will be loaded in either portrait or landscape (depending on device) but the user will never be able to switch between the two?  There are some issues when switching back and forth, but probably they have always been there.

Patch-set: 2