Fix crash when switching between LTR/RTL
When the language is changed the order of tabs might change but
the fragments stored in the fragment manager remains the same (becauase
the fragment manager persists fragments across rotation and configuration
changes).

Retrieve the fragments from the fragment manager in an RTL adjusted
fashion in instantiateItem, and also make the logic more resilient to
ClassCastExceptions.

Bug: 17673059
Change-Id: I1f9b7d85b6a4ecf3ff155c9e6e8cf875ddd51463