Update patch set 2
Patch Set 2:

Hi there, I've just seen this commit that was merged some months ago. Would it possible to allow both Dialer and Contacts to rotate to landscape mode (it can be a user setting to be toggled and it can be unchecked by default)?

A quick example: you're driving and your phone is on your car deck in landscape mode with a navigation app running (such as Google Maps, Waze, etc). If you receive a call, then the Dialer is forces the UI to portrait, which makes it difficult to read/use. The same applies if you need to Dial a number or use the Contacts app: they are all forced to portrait mode but the phone is sitting on the car deck in landscape.

Patch-set: 2
Reviewer: Ruukku Puu <9922@1ec6b3db-b2c1-4fa4-84a2-4c7efe89ba71>
Label: Verified=0