Remove accessibility delegate on shortcut card.
An accessbility delegate is set in the Call Log Adaptor to
automatically expand the call log list item when tapped in
accessibility mode. We don't want this for the shortcut card,
because it shouldn't be expandable.

When not doing this expansion, the shortcut card now obtains
accessibility focus and reading properly.

Change the description text for the shortcut card so it reads
"call back #/name". Should update this later with a string to
include the other information, but it's not straightforward to
do right now.

Bug: 17929331
Change-Id: I691fe238ea436916ab7505c39039b6c1cea27781