Update Conference Manager UI after capability changes.
Before, we just set the visibility of the disconnect and separate
buttons based on the initial call capabilities when the fragment is
created.

This can actually change, for example if an IMS call changes to
a circuit switch GMS call.

To address this, now implement a call details listener on the
conference manager presenter, and update the UI if there is a change
to disconnect or separate capabilities on a call.

This required rearranging some code in the fragment to accomodate
these changes; we now no longer assume the code to set the visibility
of the buttons is only called once.

Bug: 17429707
Change-Id: I476202a21a8ce07568170cd0e046697df84df991