Update patch set 1
Patch Set 1:

Sorry about the authorship thing. I honestly did not know how to do this as this was my first ever commit to gerrit and it was 3 in the morning. 

Since you abandoned it with requiring dependencies this seems to require a more thorough merge between shipping and the main branch in the dialer is that correct?

Patch-set: 1