Update patch set 2
Patch Set 2:

This shipping branch is used only for our factory shipping devices. At certain phases of the test & release process we need to go into code freeze on the branch.

Thanks for highlighting this one. We'll certainly try to get it in as soon as we can; but right now the branch is frozen.

Patch-set: 2