Use conference call label for IMS calls.
- Use GENERIC_CONFERENCE capability to show CDMA-scenario string
and image asset.
- Since mPrimary/mSecondary and others are class variables, don't
pass them around all the time through functions if we don't need to.
- Split out the isConference logic for setting the call cards into
its own block.
- Move logic for specifying conference string / icon into the
CallCardPresenter. This makes more sense here, and means that we
don't need to pass around a bunch of call paraemters to the fragment.

Not all these changes are strictly necessary for the change I need to
do, but I had made them originally because I had been thinking of
using the VoLTE capability. Even though now I use GENERIC capability,
I think the changes are still an improvement in terms of organization
though.

Bug: 18284408
Change-Id: I41d7825611456e9ea524db9dd2ef19c9646ab7e2