Correcting population of search box with voice search results.
The issue is the onActivityResult callback which receives the result of
the voice search is called BEFORE the dialtacts activity is resumed, and
hence the search results fragment is not properly shown (since resume has
not yet completed).

Changed so that the search results are stored until onResume, where they
are applied to the search box.

Bug: 18410823
Change-Id: I02985f722643846067678a67f2c1fda76434fbc4