On MSIM devices, include account from call log in talkback.
In TalkBack mode, when a user taps on a call log entry, it should read
back the name of the SIM/account used for the call. The string used is
"on <phoneAccount>".
Also add back periods removed in the string updates because they are
needed for verbal stopping points in TalkBack mode.

Bug:18517912
Change-Id: Iaadbd859ef980ca030f9fb60d3395ca057e8cefc