Change Log.wtf to Log.w to prevent rare crash.
Occasionally the message "Secondary info clicked but no secondary call."
will cause a crash because it is logging wtf. It's not quite clear why
this happens but is not causing any behavioral problems. Also no defined
repo steps, but it's low-risk to turn Log.wtf to Log.w which will just
log a warning instead of crash.

Bug: 18465594

Change-Id: Ied23e5b52162cf68089bd5248fbdbf9e05eefdf8