Fixing manage conference accessibility issues.
1. The "Manage Conference" button would be announced twice when it received
focus.  Moved content description to the parent FrameLayout that acts as
the button and made it focusable to ensure that the description is read
only a single time.
2. When tapping "Manage Conference" button there was no audible feedback
that the manage conference screen was opened.  Set focus to the list of
conference participants so that this list is announced when the manage
conference screen opens.
3. It was possible to change focus from the list of participants in the
manage conference screen to the call card buttons and other elements
behind the manage conference screen.  Fixed by explicitly setting the
call card fragment visibility to GONE when the conference manager is shown.

Bug: 18440532
Change-Id: I0ce4f74ccb98c5cefd6c6e75eb8ef9310bd6d7fb