Update patch set 2
Patch Set 2:

On a semi-related note, my impression is that the whole ContactBuilder class seems a little overengineered. IMHO the code would be much easier if we

- used a LookupResult class with only the required fields (name, formatted number, normalized number, number type, number label, address, city, website, photo URI, business yes/no)
- fill that class from the implementations without using intermediate classes
- have that class (de)serialize itself from JSON

I realize ContactBuilder is way more flexible than that, but the whole lookup code doesn't seem to use any of that flexibility, and the usage of calllog.ContactInfo actually impairs flexibility (it e.g. can't transmit the business yes/no flag). Thoughts on this? Am I missing something?

Patch-set: 2